# Implement SIMD optimizations for BitPacker4x on aarch64