--- a/src/bitpacker4x.rs
+++ b/src/bitpacker4x.rs
@@ -1,6 +1,9 @@
 use super::{BitPacker, UnsafeBitPacker};
 
-#[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
+#[cfg(any(
+    target_arch = "x86_64",
+    all(target_arch = "aarch64", target_endian = "little")
+))]
 use crate::Available;
 
 const BLOCK_LEN: usize = 32 * 4;
@@ -70,7 +73,7 @@ mod sse3 {
     }
 }
 
-#[cfg(any(target_arch = "aarch64"))]
+#[cfg(all(target_arch = "aarch64", target_endian = "little"))]
 mod aarch64 {
 
     use super::BLOCK_LEN;
@@ -257,7 +260,7 @@ impl BitPacker for BitPacker4x {
                 return BitPacker4x(InstructionSet::SSE3);
             }
         }
-        #[cfg(target_arch = "aarch64")]
+        #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
         {
             if aarch64::UnsafeBitPackerImpl::available() {
                 return BitPacker4x(InstructionSet::NEON);
@@ -273,7 +276,7 @@ impl BitPacker for BitPacker4x {
                 InstructionSet::SSE3 => {
                     sse3::UnsafeBitPackerImpl::compress(decompressed, compressed, num_bits)
                 }
-                #[cfg(target_arch = "aarch64")]
+                #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
                 InstructionSet::NEON => {
                     aarch64::UnsafeBitPackerImpl::compress(decompressed, compressed, num_bits)
                 }
@@ -300,7 +303,7 @@ impl BitPacker for BitPacker4x {
                     compressed,
                     num_bits,
                 ),
-                #[cfg(target_arch = "aarch64")]
+                #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
                 InstructionSet::NEON => aarch64::UnsafeBitPackerImpl::compress_sorted(
                     initial,
                     decompressed,
@@ -324,7 +327,7 @@ impl BitPacker for BitPacker4x {
                 InstructionSet::SSE3 => {
                     sse3::UnsafeBitPackerImpl::decompress(compressed, decompressed, num_bits)
                 }
-                #[cfg(target_arch = "aarch64")]
+                #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
                 InstructionSet::NEON => {
                     aarch64::UnsafeBitPackerImpl::decompress(compressed, decompressed, num_bits)
                 }
@@ -351,7 +354,7 @@ impl BitPacker for BitPacker4x {
                     decompressed,
                     num_bits,
                 ),
-                #[cfg(target_arch = "aarch64")]
+                #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
                 InstructionSet::NEON => aarch64::UnsafeBitPackerImpl::decompress_sorted(
                     initial,
                     compressed,
@@ -373,7 +376,7 @@ impl BitPacker for BitPacker4x {
             match self.0 {
                 #[cfg(target_arch = "x86_64")]
                 InstructionSet::SSE3 => sse3::UnsafeBitPackerImpl::num_bits(decompressed),
-                #[cfg(target_arch = "aarch64")]
+                #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
                 InstructionSet::NEON => aarch64::UnsafeBitPackerImpl::num_bits(decompressed),
                 InstructionSet::Scalar => scalar::UnsafeBitPackerImpl::num_bits(decompressed),
             }
@@ -387,7 +390,7 @@ impl BitPacker for BitPacker4x {
                 InstructionSet::SSE3 => {
                     sse3::UnsafeBitPackerImpl::num_bits_sorted(initial, decompressed)
                 }
-                #[cfg(target_arch = "aarch64")]
+                #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
                 InstructionSet::NEON => {
                     aarch64::UnsafeBitPackerImpl::num_bits_sorted(initial, decompressed)
                 }
@@ -399,19 +402,35 @@ impl BitPacker for BitPacker4x {
     }
 }
 
-#[cfg(target_arch = "x86_64")]
+#[cfg(any(
+    target_arch = "x86_64",
+    all(target_arch = "aarch64", target_endian = "little")
+))]
 #[cfg(test)]
 mod tests {
+    use super::scalar;
     use super::BLOCK_LEN;
-    use super::{scalar, sse3};
     use crate::tests::test_util_compatible;
     use crate::Available;
     use crate::{BitPacker, BitPacker4x};
 
+    #[cfg(target_arch = "x64_64")]
     #[test]
-    fn test_compatible() {
+    fn test_compatible_sse3() {
+        use super::sse3;
         if sse3::UnsafeBitPackerImpl::available() {
             test_util_compatible::<scalar::UnsafeBitPackerImpl, sse3::UnsafeBitPackerImpl>(
+                BLOCK_LEN,
+            );
+        }
+    }
+
+    #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
+    #[test]
+    fn test_compatible_neon() {
+        use super::aarch64;
+        if aarch64::UnsafeBitPackerImpl::available() {
+            test_util_compatible::<scalar::UnsafeBitPackerImpl, aarch64::UnsafeBitPackerImpl>(
                 BLOCK_LEN,
             );
         }