--- a/src/bitpacker8x.rs
+++ b/src/bitpacker8x.rs
@@ -25,11 +25,6 @@ mod avx2 {
         _mm256_slli_si256, _mm256_srli_si256, _mm256_sub_epi32,
     };
 
-    #[inline]
-    unsafe fn left_shift_insert_32<const N: i32>(base: DataType, to_shift: DataType) -> DataType {
-        op_or(base, left_shift_32::<N>(to_shift))
-    }
-
     #[allow(non_snake_case)]
     unsafe fn or_collapse_to_u32(accumulator: DataType) -> u32 {
         let a__b__c__d__e__f__g__h_ = accumulator;
@@ -114,19 +109,6 @@ mod scalar {
             el[5] << N,
             el[6] << N,
             el[7] << N,
-        ]
-    }
-
-    fn left_shift_insert_32<const N: i32>(base: DataType, to_shift: DataType) -> DataType {
-        [
-            base[0] | (to_shift[0] << N),
-            base[1] | (to_shift[1] << N),
-            base[2] | (to_shift[2] << N),
-            base[3] | (to_shift[3] << N),
-            base[4] | (to_shift[4] << N),
-            base[5] | (to_shift[5] << N),
-            base[6] | (to_shift[6] << N),
-            base[7] | (to_shift[7] << N),
         ]
     }