A few smaller changes to the macros:
* Pass N value for shifts as a const generic. The simd packages use a legacy const generic syntax that allows these values to be passed as regular arguments, but functions outside those packages cannot implement the same interface. Passing the values as const generic arguments may also speed up scalar implementations if the target architecture has left/right shift immediate instructions.
* Add `left_shift_insert_32` to the interface used by the implementation macros as NEON has an instruction that performs exactly this operation.

The `neon` implementation of `or_collapse_to_u32()` does not actually perform an OR, but instead takes the maximum value. This is substantially faster (a single instruction) and computing the full OR does not matter as the next instruction takes the count of leading zeros.

Benchmark figures below are collected on an M1 MacBook Air. Wins and losses are pretty modest except for compress
delta which gets way faster. Graviton hosts aren't offered as part of the AWS free tier so I haven't tried this on that yet.

A few observations about the M1 results:
* The wins in the compress-delta path appear to be a result of reducing dependencies across instructions (1 subq >> 4 sub).
* Some of these operations are already partially vectorized by the compiler. For instance, decompression paths often use 64-bit or 128-bit vector operations instead of u32 operations and `num_bits()` is entirely vectorized into 16 `ldp` instructions (256 bit load) and 31 `orrb.16b` instructions.
* The M1 processor can decode many instructions in parallel and this code is largely branchless so vectorizing to reduce instruction count does not help that much, at least in a microbenchmark.

```
BitPacker4x/decompress-1                                                                            
                        time:   [72.860 ns 72.875 ns 72.894 ns]
                        thrpt:  [17.560 Gelem/s 17.564 Gelem/s 17.568 Gelem/s]
                 change:
                        time:   [-1.7626% -1.6976% -1.6413%] (p = 0.00 < 0.05)
                        thrpt:  [+1.6687% +1.7269% +1.7942%]
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  7 (7.00%) high severe

BitPacker4x/decompress-delta-1                                                                            
                        time:   [88.717 ns 89.072 ns 89.635 ns]
                        thrpt:  [14.280 Gelem/s 14.370 Gelem/s 14.428 Gelem/s]
                 change:
                        time:   [-3.8478% -3.4636% -2.8807%] (p = 0.00 < 0.05)
                        thrpt:  [+2.9661% +3.5879% +4.0018%]
                        Performance has improved.
Found 20 outliers among 100 measurements (20.00%)
  2 (2.00%) low severe
  1 (1.00%) low mild
  17 (17.00%) high severe

BitPacker4x/compress-1  time:   [134.46 ns 134.70 ns 134.94 ns]                                   
                        thrpt:  [9.4854 Gelem/s 9.5023 Gelem/s 9.5195 Gelem/s]
                 change:
                        time:   [+0.3742% +0.6982% +1.0038%] (p = 0.00 < 0.05)
                        thrpt:  [-0.9938% -0.6934% -0.3728%]
                        Change within noise threshold.

BitPacker4x/compress-delta-1                                                                            
                        time:   [155.85 ns 156.15 ns 156.44 ns]
                        thrpt:  [8.1818 Gelem/s 8.1973 Gelem/s 8.2130 Gelem/s]
                 change:
                        time:   [-54.493% -54.397% -54.289%] (p = 0.00 < 0.05)
                        thrpt:  [+118.77% +119.28% +119.75%]
                        Performance has improved.

BitPacker4x/decompress-15                                                                            
                        time:   [106.53 ns 106.61 ns 106.70 ns]
                        thrpt:  [11.996 Gelem/s 12.006 Gelem/s 12.015 Gelem/s]
                 change:
                        time:   [-0.7996% -0.7203% -0.6392%] (p = 0.00 < 0.05)
                        thrpt:  [+0.6433% +0.7255% +0.8061%]
                        Change within noise threshold.
Found 17 outliers among 100 measurements (17.00%)
  4 (4.00%) high mild
  13 (13.00%) high severe

BitPacker4x/decompress-delta-15                                                                            
                        time:   [110.40 ns 110.48 ns 110.59 ns]
                        thrpt:  [11.574 Gelem/s 11.586 Gelem/s 11.595 Gelem/s]
                 change:
                        time:   [-8.6616% -8.4134% -8.2071%] (p = 0.00 < 0.05)
                        thrpt:  [+8.9409% +9.1863% +9.4830%]
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

BitPacker4x/compress-15 time:   [137.07 ns 137.39 ns 137.67 ns]                                    
                        thrpt:  [9.2974 Gelem/s 9.3166 Gelem/s 9.3380 Gelem/s]
                 change:
                        time:   [-5.8697% -5.6046% -5.3255%] (p = 0.00 < 0.05)
                        thrpt:  [+5.6251% +5.9373% +6.2357%]
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

BitPacker4x/compress-delta-15                                                                            
                        time:   [164.59 ns 164.88 ns 165.17 ns]
                        thrpt:  [7.7498 Gelem/s 7.7632 Gelem/s 7.7771 Gelem/s]
                 change:
                        time:   [-54.069% -53.981% -53.902%] (p = 0.00 < 0.05)
                        thrpt:  [+116.93% +117.30% +117.72%]
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

BitPacker4x/decompress-31                                                                            
                        time:   [145.16 ns 145.30 ns 145.45 ns]
                        thrpt:  [8.8004 Gelem/s 8.8094 Gelem/s 8.8181 Gelem/s]
                 change:
                        time:   [+0.1781% +0.2889% +0.3998%] (p = 0.00 < 0.05)
                        thrpt:  [-0.3982% -0.2881% -0.1778%]
                        Change within noise threshold.

BitPacker4x/decompress-delta-31                                                                            
                        time:   [97.615 ns 97.686 ns 97.765 ns]
                        thrpt:  [13.093 Gelem/s 13.103 Gelem/s 13.113 Gelem/s]
                 change:
                        time:   [+3.7266% +3.8473% +4.0079%] (p = 0.00 < 0.05)
                        thrpt:  [-3.8535% -3.7048% -3.5927%]
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

BitPacker4x/compress-31 time:   [144.11 ns 144.29 ns 144.45 ns]                                    
                        thrpt:  [8.8609 Gelem/s 8.8712 Gelem/s 8.8819 Gelem/s]
                 change:
                        time:   [-8.7062% -8.4841% -8.2847%] (p = 0.00 < 0.05)
                        thrpt:  [+9.0330% +9.2707% +9.5365%]
                        Performance has improved.

BitPacker4x/compress-delta-31                                                                            
                        time:   [151.16 ns 151.37 ns 151.58 ns]
                        thrpt:  [8.4442 Gelem/s 8.4561 Gelem/s 8.4681 Gelem/s]
                 change:
                        time:   [-42.729% -42.458% -42.269%] (p = 0.00 < 0.05)
                        thrpt:  [+73.217% +73.787% +74.610%]
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

```