@fulmicoton 2022-06-30
Thank you! I've been wanting to have NEON to run on Graviton efficiently.