@mccullocht 2022-06-30
There are a few cleanups required and I should put the benchmark results in the PR description. I'll put this up for review in the next week or so.