@PSeitz 2022-07-15
This are the results from a Graviton instance `c7g`. The results look very mixed.

```
BitPacker4x/decompress-1
                        time:   [116.80 ns 117.10 ns 117.35 ns]
                        thrpt:  [10.908 Gelem/s 10.931 Gelem/s 10.959 Gelem/s]
                 change:
                        time:   [-5.5181% -5.2683% -5.0073%] (p = 0.00 < 0.05)
                        thrpt:  [+5.2713% +5.5613% +5.8404%]
                        Performance has improved.

BitPacker4x/decompress-delta-1
                        time:   [132.88 ns 132.92 ns 132.95 ns]
                        thrpt:  [9.6275 Gelem/s 9.6301 Gelem/s 9.6325 Gelem/s]
                 change:
                        time:   [-4.9277% -4.8944% -4.8609%] (p = 0.00 < 0.05)
                        thrpt:  [+5.1092% +5.1463% +5.1832%]
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) low mild
  1 (1.00%) high mild

BitPacker4x/compress-1  time:   [281.50 ns 281.51 ns 281.53 ns]
                        thrpt:  [4.5466 Gelem/s 4.5469 Gelem/s 4.5471 Gelem/s]
                 change:
                        time:   [+15.656% +15.683% +15.709%] (p = 0.00 < 0.05)
                        thrpt:  [-13.576% -13.557% -13.537%]
                        Performance has regressed.
Found 16 outliers among 100 measurements (16.00%)
  10 (10.00%) low severe
  4 (4.00%) low mild
  2 (2.00%) high severe

BitPacker4x/compress-delta-1
                        time:   [451.54 ns 451.95 ns 452.29 ns]
                        thrpt:  [2.8300 Gelem/s 2.8322 Gelem/s 2.8348 Gelem/s]
                 change:
                        time:   [-7.5762% -7.4456% -7.3142%] (p = 0.00 < 0.05)
                        thrpt:  [+7.8914% +8.0445% +8.1972%]
                        Performance has improved.

BitPacker4x/decompress-15
                        time:   [159.04 ns 159.07 ns 159.10 ns]
                        thrpt:  [8.0454 Gelem/s 8.0469 Gelem/s 8.0484 Gelem/s]
                 change:
                        time:   [-1.4687% -1.4345% -1.4013%] (p = 0.00 < 0.05)
                        thrpt:  [+1.4212% +1.4554% +1.4905%]
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe

BitPacker4x/decompress-delta-15
                        time:   [164.18 ns 164.22 ns 164.26 ns]
                        thrpt:  [7.7927 Gelem/s 7.7945 Gelem/s 7.7963 Gelem/s]
                 change:
                        time:   [-5.3163% -5.2881% -5.2626%] (p = 0.00 < 0.05)
                        thrpt:  [+5.5550% +5.5833% +5.6148%]
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild

BitPacker4x/compress-15 time:   [300.02 ns 300.06 ns 300.10 ns]
                        thrpt:  [4.2653 Gelem/s 4.2658 Gelem/s 4.2664 Gelem/s]
                 change:
                        time:   [+27.653% +27.720% +27.809%] (p = 0.00 < 0.05)
                        thrpt:  [-21.759% -21.703% -21.662%]
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

BitPacker4x/compress-delta-15
                        time:   [473.67 ns 473.70 ns 473.73 ns]
                        thrpt:  [2.7020 Gelem/s 2.7021 Gelem/s 2.7023 Gelem/s]
                 change:
                        time:   [-7.2258% -7.0541% -6.8859%] (p = 0.00 < 0.05)
                        thrpt:  [+7.3951% +7.5895% +7.7886%]
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe

BitPacker4x/decompress-31
                        time:   [206.14 ns 206.20 ns 206.26 ns]
                        thrpt:  [6.2058 Gelem/s 6.2077 Gelem/s 6.2095 Gelem/s]
                 change:
                        time:   [-4.2839% -4.2433% -4.2039%] (p = 0.00 < 0.05)
                        thrpt:  [+4.3884% +4.4314% +4.4757%]
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe

BitPacker4x/decompress-delta-31
                        time:   [106.38 ns 106.39 ns 106.41 ns]
                        thrpt:  [12.028 Gelem/s 12.031 Gelem/s 12.033 Gelem/s]
                 change:
                        time:   [+0.1206% +0.1501% +0.1815%] (p = 0.00 < 0.05)
                        thrpt:  [-0.1812% -0.1498% -0.1205%]
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

BitPacker4x/compress-31 time:   [326.89 ns 326.91 ns 326.93 ns]
                        thrpt:  [3.9153 Gelem/s 3.9155 Gelem/s 3.9157 Gelem/s]
                 change:
                        time:   [+20.088% +20.123% +20.157%] (p = 0.00 < 0.05)
                        thrpt:  [-16.776% -16.752% -16.727%]
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

BitPacker4x/compress-delta-31
                        time:   [319.09 ns 319.13 ns 319.18 ns]
                        thrpt:  [4.0103 Gelem/s 4.0109 Gelem/s 4.0114 Gelem/s]
                 change:
                        time:   [-30.273% -30.250% -30.228%] (p = 0.00 < 0.05)
                        thrpt:  [+43.323% +43.370% +43.416%]
                        Performance has improved.
```