@mccullocht 2022-07-16
The basic shape of the graviton results is similar to m1. I'll take another pass at the `compress` routines to see if I can improve things.