@mccullocht 2022-07-18
I didn't realize that `num_bits()` was part of the `compress` benchmarks and made a small change that should improve things. PTAL.

While I was doing this I was looking at the compiler output and it seems like the compiler is doing a very good job of auto-vectorizing the code and reordering the instructions -- it doesn't seem to reorder the instructions in the manual vectorization case and we may suffer for it. I commit a [change](https://github.com/mccullocht/bitpacking/commit/af735295288fd25131b8e13582b73404e02de737) in the `[neon-mostly-scalar](https://github.com/mccullocht/bitpacking/commits/neon-mostly-scalar)` branch that actually undoes quite a bit of this PR and uses the scalar implementation for everything but the few pieces that the compiler doesn't do on its own:
* delta computation and integration
* `or_collapse_to_u32()` for `num_bits()`
* `left_shift_insert_32()`

I think this is worth checking out if graviton performance remains poor or if you'd just prefer a less invasive change. It looks a little weird because the overridden methods call load/store intrinsics but the compiler optimizes this all out all the input/output data is already in vector registers for the most part.